requirements.txt
# Основные зависимости для криптобиржи
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
aiosqlite==0.21.0
pydantic==2.11.9
pyjwt==2.10.1
bcrypt==5.0.0
httpx==0.27.2
python-multipart==0.0.9

